Repositorios en Spring Boot imcompleto10

+++ src/main/java/co/edu/uniquindio/application/dto/AccommodationDTO.java
@@ -10,7 +10,7 @@ public record AccommodationDTO(
         String address,
         Double latitude,
         Double longitude,
-        BigDecimal pricePerNight,
+        Double pricePerNight,
         Integer maxCapacity,
         String amenities,
         Double averageRating,
